:zap: Increase load time from mag feed to article
Add a properties to the Feed redux state, `items` to save the current list of shown items (mag, news or videos) from the feed request.

Then using the custom routes on the `Link` component, we fake display the article path using `as` props, so we’re still on the Mag Page, and with the query and the array of items in the redux we can load almost instantly the article.