:sparkles: Create hooks to get payment info
Instead of query IP data on each load of the website and sometimes not using them, use a hook to fetch them on demand (+ saving that on the redux in case it’s updated from somewhere else).

Took the opportunity to clean code and remove useless props or props that flow too much down the tree.