Make filter work
Create reducers and actions to hold the selected state of types and spots status, and use that to refresh the map.
But also move the carrelage’s Typings into the carrelageClient (instead of a d.ts), the reason is that enum don’t work in d.ts files, so they were unusable.

So use that already created enums instead of the FilterName one. Then currenly the index for managing the filters components is the one taking care of which filter to add or remove. Later that could be nice to just ‘toggle’ the filter then the reducers could take care of the rest.